module로 분리한 코드 main.js에서 한번에 호출
